modify: SettlementsTableKey constant 폴더로 분리
