+++ src/constants/lib/SETTLEMENTS_TABLE_KEY.ts
@@ -1,0 +1,13 @@
+const keyToLabelMap = {
+    'NO': 'NO',
+    'coupon_number': '쿠폰번호',
+    'coupon_name': '관리 쿠폰명',
+    'coupon_count': '사용 건수',
+    'discount_price': '쿠폰 할인 금액',
+    'cancel_price': '쿠폰 취소 금액',
+    'sum_price': '정산 금액',
+    'coupon_use_date': '쿠폰 적용일',
+    'complete_at': '정산 완료일'
+  };
+
+export default keyToLabelMap;
